% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfitTMLE.R
\name{crossfitTMLE}
\alias{crossfitTMLE}
\title{Estimate Average Treatment Effect (ATE) using from TMLE estimator using cross-fit algorithm (generalization 1)}
\usage{
crossfitTMLE(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  family.y = "binomial",
  learners,
  control,
  n_split,
  num_cf,
  rand_split = FALSE,
  gbounds = NULL,
  Qbounds = 5e-04,
  seed = 146,
  conf.level = 0.95,
  stat = "median"
)
}
\arguments{
\item{data}{a data frame of tibble}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable}

\item{covarsT}{a vector of names of covaraites for treatment model}

\item{covarsO}{a vector of names of covaraites for outcome model}

\item{family.y}{it is the family for outcome model. It can \code{binomial() (default)} or \code{"gaussian"}}

\item{learners}{similar as\code{SL.library()} in \code{SuperLearner} package.}

\item{control}{similar as \code{cvControl()} in \code{SuperLearner} package.}

\item{n_split}{number of splits used, default \code{n_split = 3}}

\item{num_cf}{number of repetition done. The default is 5.}

\item{rand_split}{logical value; if be FALSE \code{(default)}, discordant splits for exposure and outcome model are chosen systematically; otherwise chosen randomly.}

\item{gbounds}{value between (0,1) for truncation of predicted probabilities of exposure. The defaults are 5/sqrt(n)log(n) and 1-5/sqrt(n)log(n). See \code{tmle::tmle()} for more information.}

\item{Qbounds}{used to keep predicted probabilities of outcomes values bounded away from (0,1). The defaults are 5e-04 and 1-5e-04.}

\item{seed}{numeric value to reproduce the splits distribution}

\item{conf.level}{confidence limit for confidence interval, \code{default = 0.95}.}

\item{stat}{name of the summary statistics, "median" (default) or "mean" to calculate overall ATE from repetitions specific estimates.}
}
\value{
It return a list of two elements. The first element \code{ATE} is a tibble of the estimates.
}
\description{
Estimate Average Treatment Effect (ATE) using from TMLE estimator using cross-fit algorithm (generalization 1)
}
\examples{

# See the README file for details

sum(1:5)


}
