% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statin_data.R
\docType{data}
\name{statin_sim_data}
\alias{statin_sim_data}
\title{Description of statin_sim_data}
\format{
A data frame with 3000 rows and 7 columns.
}
\usage{
statin_sim_data
}
\description{
This dataset contains information about 3000 patients (simulated) on outcome (Y), statin use (statin), age, Low-density lipoprotein (ldl_log),
diabetes, risk score, risk score categories. The simulation details are described in Zivich and Breskin (2021).
}
\keyword{datasets}
